+++ bin/pana.dart
@@ -13,6 +13,7 @@ import 'package:io/ansi.dart';
 import 'package:io/io.dart';
 import 'package:logging/logging.dart' as log;
 import 'package:pana/pana.dart';
+import 'package:pana/src/logging.dart';
 import 'package:path/path.dart' as p;
 
 const defaultHostedUrl = 'https://pub.dev';
@@ -132,32 +133,7 @@ Future<void> main(List<String> args) async {
   }
 
   log.Logger.root.level = log.Level.ALL;
-
-  if (isJson) {
-    log.Logger.root.onRecord.listen((log) {
-      var map = <String, Object>{};
-
-      if (log.loggerName.isNotEmpty) {
-        map['logName'] = log.loggerName;
-      }
-
-      map.addAll({
-        'level': log.level.name,
-        'message': log.message,
-      });
-
-      if (log.error != null) {
-        map['error'] = log.error.toString();
-      }
-
-      if (log.stackTrace != null) {
-        map['stackTrace'] = log.stackTrace.toString();
-      }
-      stderr.writeln(json.encode(map));
-    });
-  } else {
-    log.Logger.root.onRecord.listen(_logWriter);
-  }
+  initializePanaLogging(isJson: isJson);
 
   // Docker is WEIRD
   // The SIGTERM signal sent to `docker run...` DOES propagate a signal to the
@@ -283,28 +259,6 @@ Future<void> main(List<String> args) async {
   }
 
   await subscription.cancel();
-}
-
-void _logWriter(log.LogRecord record) {
-  var wroteHeader = false;
-
-  var msg = LineSplitter.split([record.message, record.error, record.stackTrace]
-          .where((e) => e != null)
-          .join('\n'))
-      .map((l) {
-    String prefix;
-    if (wroteHeader) {
-      prefix = '';
-    } else {
-      wroteHeader = true;
-      prefix = record.level.toString();
-    }
-    return '${prefix.padRight(10)} $l';
-  }).join('\n');
-
-  overrideAnsiOutput(stderr.supportsAnsiEscapes, () {
-    stderr.writeln(darkGray.wrap(msg));
-  });
 }
 
 /// A merged stream of all signals that tell the test runner to shut down